requirements.txt
# Core dependencies
boto3>=1.26.0
click>=8.0.0
rich>=12.0.0
pyyaml>=6.0
openai>=1.0.0
anthropic>=0.7.0
requests>=2.28.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0